--- test/types.ava.js
@@ -1,16 +1,0 @@
-import {DataURL} from '../lib/types.js';
-import {asciiToUint8Array} from '../lib/utils.js';
-import test from 'ava';
-
-test('create', t => {
-  let d = new DataURL();
-  t.is(d.toString(), 'data:text/plain;charset=US-ASCII,');
-  d = new DataURL('foo');
-  t.is(d.toString(), 'data:text/plain;charset=US-ASCII,foo');
-  d = new DataURL('foo', {mediatype: ''});
-  t.is(d.toString(), 'data:text/plain;charset=US-ASCII,foo');
-  d = new DataURL('foo%ff', {encoding: 'percent', mediatype: 'text/plain;k=v'});
-  t.is(d.toString(), 'data:text/plain;k=v,foo%ff');
-  d = new DataURL(asciiToUint8Array('foo\xff'), {base64: true, mediatype: 'text/plain'});
-  t.is(d.toString(), 'data:text/plain;base64,Zm9v/w==');
-});
